build: keep elf64 headers in memory image
We need the headers so we can provide the symbols to loaded libraries.
This means we cannot use objcopy to munge the binary format to elf32-i386,
as the loader doesn't understand this format.

Instead, add a separate 32-bit only stage (in elf32-i386 format) that loads
the 64-bit image into memory as its payload.  This keeps multiboot happy and
leave our headers alone.