printf: fix precision
We captured the precision function (which either returns an inline precision
or from an argument, via *12$) by value, then tries to update it to set a
default precision.  But since we captured it already, the update had no effect.

Change the code to return the precision function and store it in the fragment
instead.