scripts: support multiple objects in running image
Since we're loading multiple objects, we need to load symbol tables for all
of them.  The base addresses are no longer static (esp. after we started
loading mulitple payloads, so we need to read them from the image).

A new command is added for this: 'osv syms'.  It will try to locate symbol
tables for anything that has been loaded at this stage.