mmu: fix populate_page() vs large pages
mmap() did not consider that a large page may be used to map an area that
it is overwriting.  This can happen if the primordial 1:1 area is remapped
(as when the payload is creating some guard pages).  The result is that
random memory is overwritten with a pte.

Fix by detecting large pages and splitting them up.