sched: per-cpu timer list
Maintain the list of active timers per-cpu, instead of globally.

This removes the need for locking (other than disabling preemption).

The clock_event interface is awkward in that the callback is global, not
per-cpu.  This requires us to have a global dispatch object that then
uses cpu::current() to pick up the current cpu and dispatch the event.  We
should probably make clock_event explicitly per-cpu in the future.