sched: load balance threads on cpus
At present, a trivial algorithm is used: wake up once in a while, look for
a less-loaded cpu, and push a waiting thread from this cpu to the other cpu.

This is very simple wrt. locking, since the waiting thread is guaranteed not
to be running, and to be on the runqueue of the load balancer thread.