sched: atomic thread state
Currently, we have an atomic _waiting state for a thread, and additional
non-atomic _on_runqueue and _terminated states.  This is problematic since
a ->wake() racing with a ->stop_wait() can cause a thread to be
simultaneously running and queued.

Fix by using a single atomic variable for all theses states.