sched: per-thread exception stack
With preemption, we may switch threads in an interrupt.  This means we can
no longer use a per-cpu exception stack, since every thread in the system
can potentially be preempted.

Switch to a per-thread exception stack instead.  This adds overhead to
the context switch path, but it is negligible (a single memory write).