sched: disable preemption in wait_until() sequences
If we're preempted in a wait_until() sequence, before the evaluation
of the predicate, the scheduler will see our status::waiting state and
not put us back on the runqueue.  This means we won't evaluate the predicate
even once, and wait for a wake() which may not come.  If the predicate happens
to evaulate to true, this is wrong, since we need to exit the wait_until()
without waiting at all.

Fix by disabling preemption across a wait_until().