TLB flush and other fixes to mmap/munmap/protect
Added the missing TLB flush to the mmap()/munmap()/mprotect() operations.
Note that currently I only do this on the processor running mmap(), which
is incorrect (I put in a TODO), but is good enough for Java's use case
(which is to do these things on startup).

Also added more tests, and fixed a bug on mmap(PROT_NONE) (mprotect(PROT_NONE)
used to work, but mmap(PROT_NONE) didn't).