fd: use a mutex in order to protect the gfdt
solving a race in fget() where retrieving the file structure and increasing
its refcount is not done atomically - a concurrent close() may free the file
structure and fhold() may fault.

fdfree() was replaced with fdclose() which removes the file from the gfdt
and then calls fdrop() - that removes the ugly double fdrop() calls.
Credit goes to Avi for that.