Handle #DE (divide error) exception
In the existing code, #PF was handled correctly (generating a SIGSEGV),
but on most other x86 hardware exceptions, we just abort()ed the kernel.

The #DE (divide error) exception should, like #PF, generate a signal
(the inappropriately-named SIGFPE), and this patch does this. Strangely,
the SPECjvm2008 benchmark depends on this behavior (I didn't check its
source code to figure out why).

To make it easier to generate other signals in the future, I abstracted
the existing function handle_segmentation_fault() into a more general
generate_signal() which is used in both #PF and #DE handling.