Fix memory leak in thread creation
Thread object creation used to leak one page for the FPU state (thanks
Avi for spotting this). Fix this (add a destructor which frees the page)
and add to the test-suite a test for checking that thread creation doesn't
leak memory - and while we're at it, also checked that alloc_page() and
malloc() at various sizes do not leak memory.