sched: fix thread not queued on runqueue
We used rbtree::insert_unique(), which only inserts if the value if unique.
But uniqueness is defined in terms of the ordering function, which compares
vruntime, not object identity.  The result is that if another thread was
queued with exactly the same vruntime, then the new thread is not queued.

Fix by using rbtree:insert_equal(), which inserts unconditionally.

Fixes the callout unit test.