xen: lay down detection basic infrastructure
Xen's information can be in a variety of MSRs. We need to test for them all
and figure out in which of them lays the informations we want.

Once we determine that, xen initalization code is ready to be executed.
This needs to run as early as possible, because all xen drivers will
make use of it one way or another.

The hypercall code is heavily inspired (aka mostly copied) from Avi's
xen tentative patch, with the 5-argument hypercall removed (delayed until
we need it)

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>