mmu: correctly mask the PAT bit in large pages
In large pages, bit 12 is used for the PAT, while in small pages or nonleaf
pages it is used for the page address.

Provide a level parameter to pte_phys(), and create a new pte_phys_nonleaf(),
to do the masking correctly.

This doesn't affect anything since we don't use the PAT; but it is not
impossible that Xen does.