Changed backtrace() implementation to allow lead detection in Java
My previous implementation of backtrace() didn't work when Java was running
(see comments in the code in this commit for an explanation why), so I
wrote a third (!) implementation of backtrace() which does work in Java,
and now leak detection can be done in both Java and our own tests .so's.

Please note that backtrace() now relies on frame pointers, which are currently
omitted in the "release" build - so it can only be used now in the "debug"
build. Seeing we only use backtrace() for leak detection, I guess this is
acceptable for now.

A reminder on how to use leak detection - the easiest thing is to add
the "--leak" command line paramter to the loader, to have it start leak
detection when entering the payload (not during OSv boot):

$ make -j4 mode=debug
$ scripts/imgedit.py setargs build/release/loader.img --leak java.so Try3
$ scripts/run -d

(this will run java/Try3.class - use whatever command line you want ;-)).

Let it run till the end, and then
$ gdb build/debug/loader.elf
(gdb) connect
(gdb) osv syms
(gdb) set logging on
(gdb) set height 0
(gdb) osv leak show

Please note that "osv leak show" can take a L-O-N-G time - even a few
minutes, as Java allocates myriads of objects that GDB will now have to
inspect. In a simple test I did, more than 18689 allocations from 1300
different call chains were alive at the end, and printing all of this takes
time (so does looking at it... eventually I'll add better tools for finding
the most suspicious blocks).