Fix signal handling when no signal handler set
Previously, when we got a signal (e.g., SIGSEGV generated by the
page-fault handler), we called the signal handler even if never
set - generating an endless loop of segfaults.

Now, correctly check if the signal handler is SIG_DFL (the default)
and if so, abort(), or if it is SIG_IGN (in this case, ignore it)
and only if it is neither, run the signal handler.

This fix makes debugging easier - no more endless loops of
handlers trashing the entire memory.