Add constructor to spinlock (in C++)
Added constructor to the "spinlock" type (#include <osv/mutex.h>) when
used from C++. This makes it easier to use this type from C++.

Marked constructor as "constexpr" to allow the object to be constructed
in compile time (e.g., a global spinlock variable can be initialized in
bss). Thanks Avi for this trick!