sched: fix preemption request while preemption disabled ignored
Currently, if we're requested to preempt while preemption is disabled, we
just ignore the request and wait until the next interrupt (which is not even
guaranteed to arrive) hits us.  This isn't the best arrangement.

Fix by latching the preemption request into a 'need_reschedule' thread local
variable, and examining it during preempt_enable().