Add "todo" directory.
Following an idea raised during our last discussion on TODOs, I committed
a new directory "todo/", and one file in it, todo/mm, with TODOs and ideas
for memory management.

I suggest that we allow free commits to this directory - with no overhead
of "review" of these informal todo items.

If we decide we don't like this format, we can easily move the content of
these files to some other tool or database or whatever.