Demonstrate a poll() bug in unix-domain socketpair.
The tst-af-local.so test supposedly tested a waiting poll() on a
unix-domain socketpair, but unfortunately the test was incorrect - the
writer thread usually (even always) manged to write to the socket before
poll() started, so poll() never actually had to wait and always returned
immediately.

But it turns out, when we really do need to wait, there *is* a bug:
poll() doesn't wake up when input arrives. So this patch fixes the *test*
to demonstrate the bug. After this fix, the tst-af-local.so test now fails.
I'll fix the bug, to make the test pass again, in a followup patch.