Fix condvar_wait() bug
condvar_wait() wrongly dropped the condvar's internal lock too early,
and accessed wr->t outside the lock, meaning that a concurrent wake()
could race with it. This bug was exposed in one of the pipe() tests.

This patch fixes this bug, by holding the internal lock throughout
the execution of condvar_wait(), dropping it temporarily only while
waiting.