lock-free mutex: change and clarify the role of depth and owner
The way "owner" and "depth" were used in lockfree::mutex was messy.
Ideally, neither should be needed if we implemented a non-recursive
mutex, but following the design of ::mutex, we (re)used "owner" also as
a marker that a thread was waken to have the lock (and it's not a
spurious wake).

After this patch, owner and depth are used in lockfree::mutex *only*
for implementing a recursive mutex, and building a non-recursive
mutex should be as simple as dropping these two variables.

In more detail:

1. "owner" is no longer used to tell a woken up thread that the wake
   wasn't spurious. Instead, zero the thread in the wait-record. This
   is a familiar idiom, which we already used a few times before.

2. "depth" isn't an atomic variable, so it should only be read by the
   same thread which set it, and this wasn't the case previously. Now,
   depth is only ever written (set to 1, incremented or decremented)
   or read by the lock-holding thread - and not the lock releasing thread.

3. "owner" needs to be an atomic variable - a non-lock-holding thread
   needs to read it and recognize it isn't holding the lock - but it
   doesn't need any special memory ordering with other variables, so
   should always be accessed with "relaxed" memory ordering.