Fix crash on use of deleted callout
When a callout is deleted, it is properly deleted from the set of
callouts, but if it was the next-in-line to run, it was also saved
in a local variable while waiting for its timer to expire, and could
be run despite being deleted.

The Shrew test HTTP server (see bug 7) exposed this bug - every once
in a while (usually very quickly) when a socket was deleted had a
crash when a callout referring to the deleted socket was run.

Thanks to Guy for finding and fixing this bug.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>