loader: skip reading disk on errors
This is not a very serious issue, but goes like this: The very simple read
method we are attempting right now in the loader, will keep reading from the
disk until we reach a pre-determined max size. However, the disk is usually
smaller than this. If this is the case, XEN dmesg logs are filled with messages
indicated that we are trying to read from invalid LBAs, to the point of making
the log useless for me.

So although the annoyance is minor, the patch itself is minor too. If nobody
opposes, I can apply it.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>