lockfree-mutex: fix tracepoints' type
The tracepoints in lfmutex.cc referred to the type "mutex" outside the
"lockfree" namespace. When LOCKFREE_MUTEX is disabled in osv/mutex.h,
"mutex" is not an alias of "lockfree::mutex" so this does not compile.

Move the tracepoints inside the namespace, so they can use "mutex"
and got the right one.