sched: set up thread::current() for new threads earlier
Currently, current() is set during the thread initialization sequence,
which means that preemption prior to that point will see the wrong current().
There's an irq_enable() there, but it's not very effective since interrupts
are only disabled in that place during early smp bringup, and it's not trivial
to disable interrupts for all new threads (we?