zfs: port vop_lookup
The way lookups work in OSv are very different from Solaris and FreeBSD
as we already get a preallocated vnode passed in.  Thus the code also
is very different and basically the vnode operation is an entirely
different implementation.  We thus don't even try to keep the code,
and also leave out support for things like case insenstive lookups
or extended attributes which might never be supported or at least
will look very different.