kthread: also allow for process creation
Xen BSD code will attempt to create processes to serve as deamons for xenstore.
We can basically emulate this by creating threads.

The only thing I am doing differently from the already existent thread creation
layer that we have, is that callers will expect a struct proc to be returned. We
had this as a stub, now I am creating a small struct with just the PID to serve
as a return placeholder. The listener processes in xenstore never dies, so I am
not implementing a deallocate routine for them