Expand BSD compile flags
Here are some compile-time definitions that will make importing of BSD code
easier. They should be straightforward and trivially true for our current
setup: x86_64 is our current arch (and when we support others we can just replace
the static -D__x86_64__ with a test that goes precisely here, it makes no sense
to support code that is not SMP, and the id string is included by pretty much
every BSD file - and I believe we have no interest in that.