mutex: avoid disabling interrupts in spinlock
We disable interrupts in spinlocks as a way of disabling preemption, since
preemption with spinlock held will lead to deadlock.  However, we don't
restore the interrupt flag to its previous state, so we end up inadvertantly
enabling interrupts if the lock was taken with interrupts disabled in the
first place.

Fix by switching to preempt_disable() instead of disabling interrupts. Since
spinlocks are only used by the debug code, it doesn't matter much.