bsd: provide correct number of ncpus
As the comment states itself, mp_ncpus should be the number of CPUs
in the system, not the maximum number of CPUs. We can currently grab
that from the smp_processors variable.

I have audited the code for current users, though, and found two places
in which it is used as the maximum number of cpus. Those are changed to
MAXCPU