sched: reduce over-eager timer rearming
If a thread schedules out before its time slice expires (due to blocking)
and is then made runnable and scheduled back in again, the scheduler will
update the preemption timer, first cancelling it (or possibly setting it
for another thread), then re-setting it for the original thread.  Usually
the new expiration time will be later than the original.

For a context switch intensive load, this causes a lot of timer re-arms,
which are fairly slow.

Reduce the amount of re-arms by remembering the last expiration time we set.
If the new expiration time comes after that time, don't bother rearming the
timer.  Instead, the expiration of the already-set timer will recalculate
the expiration time and set a new expiration time.

This reduces timer arming greatly, and speeds up context switches back to
their performance before the preemption patches.