libc: implement signal()
Implement the signal() function. This is hardly a useful function in OSV,
first because our signal support is pretty broken, and second because
sigaction() is a much more portable API that should always be preferred.

Nevertheless, memcached uses signal() (to catch SIGINT, which it will
never get in OSV...), so let's implement it for the sake of completeness.