Fix msleep() timeout bug
When msleep() is woken up by a timeout and not wakeup(), it would leave
its wait record - a pointer to a structure on the stack - in _evlist.
When wakeup() or wakeup_one() is next called, it can find this structure
on the stack - and it now points to random garbage (or even unmapped
area, if the thread exited).

This bug caused an easy to reproduce crash in memcached.

The fix here is to, in case of timeout, reacquire the lock and remove this
thread from _evlist - if it still there (it might no longer be there,
if the timeout and wakeup raced).