sched: more reasonable initial thread vruntime
Currently we initialize a new thread's vruntime to the clock time.  However,
as only acquire vruntime as they run, while the clock always runs, this is
unreasonably high.

Initialize it instead to the parent thread's vruntime.  Since the parent thread
is running now, its vruntime represents fairly high priority; we may want to
tune that later.

Signed-off-by: Avi Kivity <avi@cloudius-systems.com>