Add assertions to lockfree::mutex::unlock()
assert() that unlock() is only done on a mutex locked by this thread.
This helped catch a couple of nasty lockfree-mutex bugs (see next
patches), so I think it's good to have such a check, at least for
now.

I previously avoided this assertion thinking it will significantly
impact performance, but it seems it doesn't have a big impact (at
least in benchmarks I tried). In the future we should perhaps have
a mechanism to disable assertions (or some of them) in release
builds.