Fix typo bug in lockfree::mutex::try_lock()
From the "how could I have not seen this before" department, we bring you
the following bug:

try_lock() as a last attempt tries to get a handoff from a concurrent
unlock() to take the lock for itself. Obviously, it should only do so
if the "handoff" variable is set (non-zero). The typo was that we checked
!handoff instead of handoff, so we took a non-existant handoff and
possibly got the lock while someone else was still holding it.

This bug is especially visible in Java, which uses try_lock() a lot
in its monitor implementation (it spins with try_lock() and only
uses lock() after giving up the spinning) - but wasn't visible in
my tst-mutex.cc which didn't test try_lock().