mempool: make the early allocator not depend on mempools
With dynamic percpu allocations, the allocator won't be available until
the first cpu is created.  This creates a circular dependency, since the
first cpu itself needs to be allocated.

Use a simple and wasteful allocator in that time until we're ready.  Objects
allocated by the simple allocator are marked by having a page offset of 8.