Condvar: Remember the mutex that the user associated with a condvar
Until now, we allowed each condvar->wait() call to specify a different
mutex. To support "wait morphing", we need the condvar's waker to know
which mutex the waiter wants to lock; If it can be a different mutex
for each wait, we'll need to remember the mutex in wait_record. Adding
a field to condvar's wait_record but not to mutex's wait_record is possible
but quite ugly. This patch implements a simpler solution:

In practice, condvar users normally "associate" a single mutex with a
condvar, and use just it when wait()ing on the condvar. Posix threads
even officially supports only this use case, and pthread_cond_wait(3p)
states that "using more than one mutex for concurrent ... pthread_cond_wait
operations on the same condition variable is undefined".

So this patch remembers for a condvar the last mutex used in condvar->wait(),
and in a later patch we will use it to implement wait morphing: a wake()
will take this mutex, instead of waking up the thread to take it. We add
assertions that verify that this assumption is not broken by the user.

The price we pay for this simplicity is the new assumption on the single
mutex per condvar, and adding 8 more bytes to the size of a condvar.