Condvar: Do not hold the condvar's internal lock during a wake_all()
During a wake_all() of a large number of threads, we hold the condvar's
internal lock throughout the loop of waking up the threads. The reason we
need this lock is timeouts: when a condvar_wait times out, it needs to
remove its wait_record from the queue, and it can potentially happen
concurrently with a wake_all(), so we don't want items being deleted from
the list while we're scanning it.

This patch protects against this concurrency in a different way, not
requiring to hold the internal mutex for a long period of time. Under
the lock, it just grabs the whole wait list (copies the head and resets
it to zero), and then walks this list and wakes the wait_records on it
without the lock. When a timeout occurs, condvar_wait() grabs the lock
and looks for its wait record in the queue; If this happened before
condvar_wake_all() zeroed the list, it will succeed in removing itself
from the list. But if the wait_record is not on the list, it means
condvar_wake_all() already zeroed the list, but we don't know if it
already woke the wait_record, or just about to do so, so we can't
return from condvar_wait() just yet - we need to wait() on this
wait_record. This wait() will either return immediately, or wait until
the wake loop in condvar_wake_all() gets to this record and the lock
we need is taken for us.

The idea is that we improve a common case (waking many threads) and
only suffer (the potential added wait) in a rare case of a race between
a wake and a timeout.

Despite these advantages, it is dubious how helpful this patch is in
real scenarios. First of all, holding the internal mutex during a wake
only hurts when the woken up threads access the condvar, and this usually
happens only if the work done by the woken-up thread is so trivial that
it immediately gets to condvar->wait() again - and it's not clear that this
is an interesting scenario.
Second, if the wake_all() is anyway called with the user mutex locked
(as it is often does in our code), none of the woken threads can actually
wake up until wake_all() is done, so no code will try to touch the condvar
while it is locked for the entire duration of wake_all().
It seems that this patch is only likely to help in artificial benchmarks
like glibc's cond-perf.c