dhcp: networking stack hook for dhcp on the rx path
allows the dhcp worker to see incoming dhcp packets using a simple hook.
no reason to make this kind of hook generic (a BPF alternative) as there
are already many protocol specific hooks in the networking stack.

if we ever need to do more hooking, the pfil interface may be used or we
can implement a generic solution, but it's not reasonable to do so for
a single user (dhcp).