bsd: initialize taskqueue
Xenbus will use BSD taskqueues in the following way:

  taskqueue_enqueue(taskqueue_thread, &xbs->xbs_probe_children);

taskqueue_thread seems to be the main bsd taskqueue, which wasn't ported in our
initial port. So I am definiting it and initializing.

It seems a bit odd to do this in net.cc, but this is where the rest of the BSD
initialization lives anyway (maybe the filename should change?). Let me know if
you feel strongly about it, and I will change. But for me it is okay for now.