bio.h: augment bio structures for xen blkfront
There are many other bio fields and functions that blkfront uses that are
not provided by our simplified bio. I am adding them here. One of the things we need is
a type definition of daddr_t. This already exists in fatfs.h, but it makes more sense to
live in the lower level code. So instead of including fatfs in bio.h, I am including
bio.h in fatfs and moving the daddr_t definition here.