BSDEDIT/blkfront: implement disk_alloc
Disk alloc is a simple function that, well, allocates a disk. It is supposed
to be provided by the device code, but for simplicity I am providing it in
blkfront itself. We have no really better way to put it: device.c is quite generic,
and we seem to have no disk specific file.

If time brings us more users of such function, then I will move it somewhere else.