xen fixes for build.mk
We need bsd/ directory listed in includes not only for bsd/ files, but
for our cpp code as well. So it has to be in the main includes.

Second, I was overly optmistic about the interface version. The drivers
expect an inteface version slightly older than the maximum present in headers.