speed up pci discovery
We are currently using PCI's brute force enumeration, which consists of scanning
all possible combination of b,d,f. This makes us issue many inb operations, a lot
more than we could do if we were using smarter enumeration. In my system, the inbs
alone account for more than 5M cycles, which takes around 4s to run (remember this
is xen, it is faster in KVM)

We can do a lot better by using recursive discovery, and only scanning the buses
we need to. In the case of virtual systems the difference is even bigger, since it
is very unlikely that we will have any complex setup with more than one bridge.

Boot for me feels instantaneous now (once we get to start OSV)