msleep: respect BSD's PDROP
So after all the BSD code is not buggy, it is just their semantics that is
slightly different from our version of mlock (Thanks Christoph). Whether or not
we will drop the lock will be controlled by the value of the PDROP flag.

Our msleep queues are protected by an internal lock which is not the same lock
the user passed to the msleep call. Therefore, we can just a version of
wait_until that does not take a lock argument, and do the locking manually
ourselves. We may then lock it back or not, depending on the presence of the
PDROP flag.