Allow device communcation of the maximum io size
Some block devices, such as the Xen blkfront, will have a maximum per-request size
that needs to be obeyed. In order for us to do it, we'll need to be aware of this
requirement in upper layers. So set up the device with it. We initialize the default
structure to make sure only interested parties need to set this up.