console: automatically insert CR
Our console works well with KVM, that redirects the serial to stdio. In Xen,
this seems not to work very well (some documentation sources - not all - lists
'serial="stdio"' as a valid option, but it doesn't really work). In that case,
LFs are not automatically translated into CRLF and the terminal output is totally
borked.

This is obviously not exclusive to Xen: by changing our qemu command line to run
a pts instead of stdio, and then connecting to that pts, the same thing happens.

This patch modifies our console's write() function to respect the termios ONLCR
flag, which is now set by default. When write() sees a newline character and
this flag is set we output a CF before any LF.

The termios structure is left outside the specific console, and we pass a
pointer down to it. Our pre-console implementation of simple_write outputs CRs
unconditionally.

The normal stdio serial still works. Telnet still works.