BSDEDIT/blkfront: fix sector number calculation
BSD expects its sector number to be already provided by the bio. We could add this field
to the bio, but it is easier just to calculate it from the given offset.

There are many places in which the bios are filled up, including many in the zfs code. So
it is easier to change them just here.