fix idiotic xen stack corruption for the blkfront driver
While planning to run tests on Xen today, I found my guests in current tip
failing to mount ZFS. I spent some time debugging the memory allocator, since
it was the culprit last time: only to find out we were not even reaching the
memory allocator.

I noticed then that ZFS was failing with error 75 -> EOVERFLOW. Looking
further, one of our bootup messages was showing the disks as "0MB".

That information is read from the xenstore, and it was being read correctly.
However, by the time we calculate the disk size, this is no longer correct,
indicating a stack corruption.

I found out the culprit to be a subsequent call to xs_gather, which calls a
variant of scanf internally. The call was being passed a %lu argument to read
an int variable, which would explain the corruption if the # of sectors was
right before it in the stack.

Indeed, with this fix, ZFS fails in a different way now =)