__fxstat64: Don't check version argument
In Linux, _STAT_VER is 1 on 64-bit (and 3 on 32-bit), but glibc never
verifies the argument to __fxstat64. JNR - a library used by JRuby -
wrongly (I believe) passes ver==0 to __fxstat64
(see jnr-posix/..../LinuxPosix.java). On Linux this wrong argument is
ignored but in our implementation, fails the check.

So this patch removes this check from our code as well, to let JNR and
therefore JRuby which uses it, use stat without failing.