run.py: also execute xen
So far, our run.py script only runs qemu-based VMs, like KVM. For Xen, I was
resorting to manually run xl. This patch implements xen support in that script.

We do need a bit of extra information, like the bridge name. I am defaulting to
F19's, allowing the user to set it up. We can do better by trying to guess from
reading what is in the system, but this will do for now.

Another problem is the console, which Xen does not redirect to stdio. We have
to attach one, but detaching it does not kill the guest, as it happens with the
qemu-based VMs. Because of that, I am defaulting to include the "-c" option to
automatically attach a console, but for people who would like to have it on
background, an option to not include it.