initialize console later
Some time ago I have moved the console initialization a bit earlier, so
messages could be seen earlier. This has been, however, creating spurious
problems (1 at each 10 - 15 boots) on Xen HVM. The reason is that the isa
serial reset code enables interrupts upon reset, and the isa irq interrupt
will call wake() in the pool thread, which at this point is not yet started.

Since these days we already have simple_write() dealing with the early stuff,
move it back to where it used to be.

P.S: Dima found a way to make this problem 100 % reproduceable, by queueing
data in the input line before the console starts. With this patch, the problem
is gone even if Dima's method is used.