only run dhcp if we have a network interface
As I have stated previously, what is true for qemu (that we always have
a user-provided network interface) is not true for Xen. It is quite possible
that we boot with no network interface at all. In that case, we will get stuck
asking for an IP that will never come.

This patch takes care to call for dhcp only if our interface is really up. Since
networking is such a core service, we'll print a message if we can't do that.