Add copyright statements in libc/
Add Cloudius copyright statements in libc/.

One unfortunate problem we have in libc/ is that most of the code there was copied
from Musl, and Musl is copyrighted and licensed, but individual files do not carry
any statements - Rich Felker explains in Musl's COPYRIGHT file: "The decision to
exclude such comments is intentional, as it should be possible to carry around the
complete source code on tiny storage media."

In the future, we should find all the Musl files and add to each one the line
Copyright (C) 2005-2013 Rich Felker. But we'll need to do it carefully so as not
to wrongly attribute non-Musl code to Musl. In the meantime, if we just have
documentation/LICENSE-musl, and don't claim our own copyright on the musl files,
I think we're playing fair.