vfs_task: Fix path conversion in certain cases
task_conv was wrongly converting "/.." and produced a path with unexpected
characters after /
Consequently, repeated calls (Java) to 'new File("/..").exists()' returned
ambigous results, not to mention other properties like isDirectory.

This confused the shell's completion mechanism (and me).

In more detail, in this (/..) situation 'len' was more than 2 (in line 105) which
caused 'tgt' to be pointing before the begining of 'full', from that point
whatever stopped the backward search for '/' (lines 109-112) was unexpected.
I guess this might occur with other combinations.

This sets 'len' to be 1 when starting an absolute path conversion, to match
the behavior where relative path conversion sets 'len = strlen(cwd)'.

Signed-off-by: Or Cohen <orc@fewbytes.com>