Trivial prototype change
Change the name of the argument of wake_with from "Pred" to "Action".
This argument is a function to run, *not* a predicate (it's not supposed
to return a boolean value), so it doesn't make sense to call it Pred.

The implementation of wake_with already used the name "Action" - this
patch fixes the prototype too.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>