condvar: add comment
Add a comment to condvar explaining that it makes a guarantee that
POSIX Threads' condition variables do not - that there are no spurious
wakeups.

The comment goes on to explain that at least in one place (semaphore.cc)
we make use of this added guarantee, so if the condition variable
implementation is ever rewritten, we'll need to keep this guarantee.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>