tests: Allow specify test names to testrunner
Allow specify test names to testrunner such as testrunner.so tst-foo1.so
tst-foo2.so to run tst-foo1 and tst-foo2 only in the order they are
specified. This does not change the default behavior that
cmdline=testrunner.so will run all tests under /tests dir.

This change can help developer to run the only test cases they need
quickly to verify new feature or find possible regressions.

Signed-off-by: Yang Bai <hamo.by@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>