Allow using floats and double values in tracepoints
The new scheduling algorithm is based on floating point runtimes,
and I'd like to be able to tracepoints to debug it.

Adding floats and doubles is trivial - we just need to tell the C++
tracing code to use the character "f" for floats and "d" for doubles,
and Python (run in gdb when we use "osv trace") already knows how to unpack
values with these designations.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>