Change default signal handler to do poweroff() instead of abort()
Our default signal handler (for, e.g., ^C) used abort(), which results
in a hung VM, and is not very useful. Change this to osv::poweroff(),
which gives more useful behavior. Now one can, for example, run a
test with run.py and interrupt it in the middle with ^C causing the
VM to power off and run.py to complete.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>