Fix make mode=debug build
When building with -O0 we get:

 bsd/sys/cddl/contrib/opensolaris/uts/common/fs/zfs/zfs_ioctl.o: In
 function `zfs_ioc_destroy_snaps_nvl':
 (...)/bsd/sys/cddl/contrib/opensolaris/uts/common/fs/zfs/zfs_ioctl.c:3265:
 undefined reference to `zvol_remove_minor'

Looks like the problem is that the missing symbol,
`zvol_remove_minor', referenced from zfs_ioctl.o is not in the zvol.o
because it's not defined, because it is inside '#ifdef NOTYET' block
(which is off). It works in the normal build because when compiled
with -O2 (and even with -O1), the reference to zvol_remove_minor in
zfs_ioctl.c is optimized away.

This patch puts the block using zvol_remove_minor also inside '#ifdef
NOTYET' to be consistent.

Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>