Simplify host-side of zfs image build
This patch simplifies the host-side work in the new /usr zfs filesystem
build process.

Previously, we copied the files to a temporary directory, used "cpio"
to archive them and sent its output to the guest with "netcat".

With this patch, we no longer have a temporary directory, and do not
need either cpio or netcat on the build machine.

Rather, mkzfs.py itself, using python (instead of a separate "nc" process),
connects to the guest and sends it the files - still using the CPIO format.

Rather than arbitrarily sleep for 3 seconds before the host tries to
connect to the guest (which might not be enough for some, or a waste
of time for others), with this patch the host looks at the guest's output
and connects when it sees the message "Waiting for connection".

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>