Add /tmp entry to usr.manifest
When this line is absent in usr.manifest:

  /tmp/index.rb: ../../mgmt/web/views/index.rb

... tst-rename.so test fails with:

  Assertion failed: c_str (<...>/osv/external/misc.bin/usr/
    include/boost/filesystem/path_traits.hpp: dispatch: 214)

The reason for failure is missing /tmp. It is supposed to be created
in unpack_bootfs() but it is later masked by zfs root pivot. The line
in the manifest recreated it, but since mgmt can be a separate module
it may not always be there.

This patch fixes that by adding an explicit entry for /tmp to
usr.manifest, same as we do for /dev.

Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>