Remove global variable usage from run.py script
Previously, the variables image_file, cmd_args and opt_path were used globally
along the run.py script, I did not consider this convenient and thus, opted to
convert these global variables, into one variable that is passed as
a parameter.

This variables was renamed to options, with the mindset that, if desired, this
variable could come from a configuration file, thus making the passing of
command line arguments optional. So, if this functionality for a configuration
file... Or perhaps default values, be added to the script, there would not be
much need for refactoring (or renaming) as there would with these global
variables.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@gmail.com>